requirements.txt
sqlalchemy==2.0.35
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pypdf==5.0.1
sentence-transformers==3.1.1
langdetect==1.0.9
requests==2.32.3
cryptography==43.0.1